fix(run): Fix multiple launch when run at same time
Without re-affecting the scheduled command, the execute command was done but the entity wasn't marked as done.
In a result, when 3 command must be launch at same times, the second one and the third one were launch multiple times.
The fault was to clear the entityManager, that came from very old days when we didn't use subprocess to launch commands.
In addition, in case connection is lost, we kept the find(id) to force the reconnection